chore(seo): clean up #58 slice A — move helper out of __tests__/, dry test, trim script
Post-merge cleanup of code-quality issues in PR #59.

- Move extractNAP from `src/__tests__/helpers/extractNAP.ts` to
  `src/utils/extractNAP.ts`, drop the inline `describe()` block and
  `eslint-disable` comment that worked around Jest treating files in
  `__tests__/` as test files.
- Update napByteMatch.test.ts import path; collapse three near-identical
  test bodies into a single `test.each(surfaces)` table.
- Trim check-forbidden-strings.sh from 81 to ~30 lines (kept the
  `Lowry`/`Lowrey` distinction comment and the rc==0|1|2 handling).

114/114 Jest tests pass locally; build:prod green; check-forbidden-strings.sh exit 0.
Net: -245 / +80 lines (cleanup file edits only; .delivery/memory not in this PR).

Bezalel: lighter pattern, same shape on the mountain.
